feat: add newItemsExtractorService to extract new items
