+++ apps/api/src/items-generator/steps/data-aggregation/index.ts
@@ -1,0 +1,4 @@
+export { SharedUtilsService } from './shared-utils.service';
+export { NewItemsExtractorService } from './new-items-extractor.service';
+export { AiDeduplicatorService } from './ai-deduplicator.service';
+export * from './prompts.constants';
